style(domain service) lint and formating

+++ services/domain/src/bus/handlers/user_deleted_handler.ts
@@ -1,18 +1,27 @@
-import { type EventPayload } from "@services/eventbus";
-import { ProfileModelInstance, DomainModelInstance, UserRoleModelInstance } from "../../models";
+import type { EventPayload } from "@services/eventbus";
+import {
+	DomainModelInstance,
+	ProfileModelInstance,
+	UserRoleModelInstance,
+} from "../../models";
 
-export default async function HandleUserDeleted(message: EventPayload): Promise<void> {
-    try {
-        const { id } = message!.message;
-        if (!id) {
-            throw new Error("Missing required fields in message");
-        }
-        await DomainModelInstance.multiTableTransaction(async (conn) => {
-            await DomainModelInstance.deleteByOwnerId(id, conn)
-            await ProfileModelInstance.delete(id, conn)
-            await UserRoleModelInstance.deleteByUserId(id, conn)
-        });
-    } catch (err) {
-        throw new Error("Error handling USER_DELETED event: ", { cause: err });
-    }
+export default async function HandleUserDeleted(
+	message: EventPayload,
+): Promise<void> {
+	try {
+		if (!message) {
+			throw new Error("Missing message in event payload");
+		}
+		const { id } = message.message;
+		if (!id) {
+			throw new Error("Missing required fields in message");
+		}
+		await DomainModelInstance.multiTableTransaction(async (conn) => {
+			await DomainModelInstance.deleteByOwnerId(id, conn);
+			await ProfileModelInstance.delete(id, conn);
+			await UserRoleModelInstance.deleteByUserId(id, conn);
+		});
+	} catch (err) {
+		throw new Error("Error handling USER_DELETED event: ", { cause: err });
+	}
 }
